chore: cleanup sponsor lp components
